requirements.txt
Flask~=2.0.3
pycryptodome
requests
python-crontab
psutil
pyinstaller
